Set empty Series dtype and tidy examples
Set the empty pandas Series in CompletenessIndex to dtype=float to avoid ambiguous/deprecated dtype behavior. Clean up example/main blocks: switch example ERA5 instantiation to non-downloading mode and comment out retrieval/store calls (and add a commented MeteoRaster plot example) to avoid accidental network/file ops; add a commented ALARO T2M instantiation and capture read_local into `mr` in the IRM examples. These are non-functional/example changes to reduce unintended runs during development.

+++ tethys_tasks/base.py
@@ -21,7 +21,7 @@ class CompletenessIndex():
 
         self.folder.mkdir(exist_ok=True, parents=True)
         
-        self.index = pd.Series([], index=pd.Index([], name='file_name'), name='complete')
+        self.index = pd.Series([], index=pd.Index([], name='file_name'), name='complete', dtype=float)
 
         self.read()
         self.check_existance()
